requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
entrypoints==0.3
flake8==3.7.9
google-api-core==1.16.0
google-auth==1.12.0
google-cloud-texttospeech==1.0.1
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
mccabe==0.6.1
mutagen==1.44.0
pkg-resources==0.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycodestyle==2.5.0
pydub==0.23.1
pyflakes==2.1.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
